requirements.txt
torch
torchvision
transformers
Pillow
numpy
scikit-learn
